wraped up attach/detach modules to roles

+++ app/Http/Controllers/Api/RoleController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\Api\RoleRequest\ReplaceRoleRequest;
 use App\Http\Requests\Api\RoleRequest\StoreRoleRequest;
 use App\Http\Requests\Api\RoleRequest\UpdateRoleRequest;
 use App\Http\Resources\Api\RoleResource;
+use App\Models\Module;
+use App\Models\Permission;
 use App\Models\Role;
 use App\Traits\ApiResponses;
 use Illuminate\Auth\Access\AuthorizationException;
@@ -134,48 +136,79 @@ class RoleController extends ApiController
             return $this->error('You are not authorized to delete a Role.', 401);
         }
     }
-    
-    public function attachModule(Request $request, string $uuid)
+
+    public function attach(Request $request, string $uuid) 
     {
-        try {
-            $role = Role::where('uuid', $uuid)->firstOrFail();
-            $modules = $request->input('modules');
+        $role = Role::where('uuid', $uuid)->firstOrFail();
+        $currentModules = [];
+        foreach ($role->modules as $module)
+            $currentModules[] = $module->uuid;
+        
+        $inputModules = $request->input('modules');
+        $toUpdate = $toCreate = [];
 
-            foreach ($modules as $module) {
-                $role->modules()->attach($module['id'], [
-                    'is_read' => $module['isRead'],
-                    'is_write' => $module['isWrite'],
-                    'is_delete' => $module['isDelete'],
-                    'is_active' => $module['isActive'],
-                    'created_at' => now(),
-                    'updated_at' => now()
-                ]);
+        foreach ($inputModules as $input) {
+            if (in_array($input['uuid'], $currentModules)) {
+                // update existing
+                $toUpdate[$input['uuid']] =  [
+                    'is_read' => $input['isRead'],
+                    'is_write' => $input['isWrite'],
+                    'is_delete' => $input['isDelete'],
+                ];
+                
+            } else {
+                try {
+                    // confirm existence
+                    $module = Module::where('uuid', $input['uuid'])->firstOrFail();
+                    // create new
+                    $toCreate[$module['id']] = [
+                        'is_read' => $input['isRead'],
+                        'is_write' => $input['isWrite'],
+                        'is_delete' => $input['isDelete'],
+                        'is_active' => $input['isActive'],
+                        'created_at' => now(),
+                        'updated_at' => now()
+                    ];
+
+                } catch (ModelNotFoundException $ex) {
+                    return $this->error('Module does not exist.', 404);
+                }
             }
-
-            return RoleResource::collection(
-                Role::with('modules')->where('uuid', $uuid)->get()
-            );
-
-        } catch (ModelNotFoundException $ex) {
-            return $this->error('Role or Module does not exist.', 404);
         }
+
+        if (!empty($toUpdate))
+            foreach ($role->modules as $module)
+                if (isset($toUpdate[$module->uuid]))
+                    $module->pivot->update($toUpdate[$module->uuid]);
+
+        if (!empty($toCreate))
+            foreach ($toCreate as $id => $attributes)
+                $role->modules()->attach($id, $attributes);
+        
+        return new RoleResource($role->load('modules'));
     }
 
-    public function detachModule(Request $request, string $uuid)
+    public function detach(Request $request, string $uuid)
     {
         try {
             $role = Role::where('uuid', $uuid)->firstOrFail();
-            $modules = $request->input('modules');
+            $currentModules = [];
+            foreach ($role->modules as $module)
+                $currentModules[] = $module->uuid;
 
-            foreach ($modules as $module) {
-                $role->modules()->detach($module['id']);
+            $inputModules = $request->input('modules');
+            foreach ($inputModules as $input) {
+                // confirm existence
+                $module = Module::where('uuid', $input['uuid'])->firstOrFail();
+                // remove from role
+                if (in_array($module['uuid'], $currentModules))
+                    $role->modules()->detach($module['id']);
             }
-
-            return $this->ok("Detach module/s is successful", []);
+            
+            return new RoleResource($role->load('modules'));
 
         } catch (ModelNotFoundException $ex) {
             return $this->error('Role or Module does not exist.', 404);
-        }
-        
+        }   
     }
 }
